+++ web/usuaris/src/__tests__/main.test.tsx
@@ -2,6 +2,8 @@ import { render, screen } from '@testing-library/react'
 import { RouterProvider, createMemoryRouter } from 'react-router-dom'
 import { createGlobalStyle } from 'styled-components'
 import { font } from '@itacademy/ui'
+import { QueryClientProvider } from '@tanstack/react-query'
+import { queryClient } from './setup'
 import { ErrorPage, Home } from '../pages'
 import { Layout } from '../components/layout'
 
@@ -26,10 +28,12 @@ describe('main', () => {
     })
 
     render(
-      <Layout>
-        <GlobalStyle />
-        <RouterProvider router={router} />
-      </Layout>
+      <QueryClientProvider client={queryClient}>
+        <Layout>
+          <GlobalStyle />
+          <RouterProvider router={router} />
+        </Layout>
+      </QueryClientProvider>
     )
 
     const HomeElement = screen.getByText('Configuración')