+++ web/usuaris/src/constants/index.ts
@@ -1,0 +1,4 @@
+import { urls } from './urls'
+import { paths } from './paths'
+
+export { paths, urls }