+++ web/usuaris/src/main.tsx
@@ -3,6 +3,7 @@ import 'modern-normalize/modern-normalize.css'
 import { createGlobalStyle } from 'styled-components'
 import React from 'react'
 import { RouterProvider, createBrowserRouter } from 'react-router-dom'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { font } from '@itacademy/ui'
 import { paths } from './constants/paths'
 import { ErrorPage, Home } from './pages'
@@ -22,11 +23,15 @@ const router = createBrowserRouter([
   },
 ])
 
+const queryClient = new QueryClient()
+
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <Layout>
-      <GlobalStyle />
-      <RouterProvider router={router} />
-    </Layout>
+    <QueryClientProvider client={queryClient}>
+      <Layout>
+        <GlobalStyle />
+        <RouterProvider router={router} />
+      </Layout>
+    </QueryClientProvider>
   </React.StrictMode>
 )