+++ web/usuaris/src/types/index.tsx
@@ -1,0 +1,1 @@
+export { type TUserData } from './types'